requirements.txt
streamlit==1.51.0
streamlit-autorefresh==1.0.1
pandas==2.2.2
numpy==1.26.4
plotly==5.9.0
requests==2.32.3
supabase-python